Added sync/async ram_infer output selection
The output flops for the ram_infer module are now selectable by means
of a parameter, SYNC_OUTPUT. Output flops are enabled by default. A
third option, though I can't think of a usage scenario, is to allow
for synthesis of latches gated by the input enable signal at the
output.