Changed asynchronous mode to include latch
This was giving me synthesis problems where Xilinx would infer a TON
of flops instead of actual Block RAMs. I think this problem originated
from the lack of an output latch, as Xilinx documentation seems to
indicate that the outputs from BRAM elements always go through a latch
and may optionally go through a flop. Forcing the design to not use a
latch should result in ISE inferring flops for everything. All,
theoretical of course...