Eye centering of uart_rx and bug fix for tx
The read using `uart_rx.v` should now be better centered in the eye of
the rx bit. This may have zero impact unless the intended clock rate
is some small multiple of the UART clock rate.

`uart_tx.v` was allowing the data buffer to be written for longer than
it should have been. This is now limited to one cycle.