Reorder sqrt_generic reset for synthesis
The reset ordering, while efficient (was thrown at the end of the always
block instead of burning a forcing an "else begin/end" block) wasn't
supported by Cadence RTL Compiler 14.1. I changed this so that this
now synthesizes with RC 14.1.

Added the do script for sqrt_generic to create a work library.