Properly apply Apache v2 License
This adds an actual Apache v2 License as opposed to the truncated text.
All src/sim files are then modified to use the correct version of the
Apache text (and not the lazy "See License" approach).

No functional change to the licensing, just making it more explicit as to
what is going on.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>